+++ src/main/java/fr/gamecreep/basichomes/commands/create/CreateHome.java
@@ -1,12 +1,12 @@
-package fr.gamecreep.basichomes.entities.commands.teleport;
+package fr.gamecreep.basichomes.commands.create;
 
 import fr.gamecreep.basichomes.BasicHomes;
 import fr.gamecreep.basichomes.entities.enums.Permission;
 import fr.gamecreep.basichomes.entities.enums.PositionType;
 
-public class TeleportHome extends TeleportCommand {
+public class CreateHome extends CreateCommand {
 
-    public TeleportHome(BasicHomes plugin) {
+    public CreateHome(BasicHomes plugin) {
         super(plugin, PositionType.HOME, Permission.USE_HOME);
     }
 }
