+++ src/main/java/fr/gamecreep/basichomes/commands/delete/DeleteCommand.java
@@ -1,4 +1,4 @@
-package fr.gamecreep.basichomes.entities.commands.teleport;
+package fr.gamecreep.basichomes.commands.delete;
 
 import fr.gamecreep.basichomes.BasicHomes;
 import fr.gamecreep.basichomes.Constants;
@@ -7,7 +7,6 @@ import fr.gamecreep.basichomes.entities.enums.Permission;
 import fr.gamecreep.basichomes.entities.enums.PositionType;
 import fr.gamecreep.basichomes.files.DataHandler;
 import lombok.NonNull;
-import org.bukkit.Location;
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
@@ -18,13 +17,13 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public abstract class TeleportCommand implements CommandExecutor, TabCompleter {
+public abstract class DeleteCommand implements CommandExecutor, TabCompleter {
     private final BasicHomes plugin;
     private final PositionType type;
     private final Permission permission;
     private final DataHandler handler;
 
-    protected TeleportCommand(BasicHomes plugin, PositionType type, Permission permission) {
+    protected DeleteCommand(BasicHomes plugin, PositionType type, Permission permission) {
         this.plugin = plugin;
         this.type = type;
         this.permission = permission;
@@ -42,25 +41,28 @@ public abstract class TeleportCommand implements CommandExecutor, TabCompleter {
                 return true;
             }
 
-            if (args.length == 0) {
-                this.plugin.getChatUtils().sendPlayerError(playerSender, String.format("Please add the name of the %s to teleport to !", this.type.getDisplayName()));
-                return true;
+            if (args.length < 1) {
+                this.plugin.getChatUtils().sendPlayerError(playerSender, String.format("Please add the name of the %s to delete !", this.type.getDisplayName()));
+                return false;
             }
 
             String name = args[0];
-            SavedPosition pos = this.handler.getByName(playerSender, name);
+            SavedPosition pos = this.plugin.getHomeHandler().getByName(playerSender, name);
 
             if (pos == null) {
                 this.plugin.getChatUtils().sendPlayerError(playerSender, String.format("No %s exists with that name !", this.type.getDisplayName()));
                 return true;
             }
 
-            Location location = pos.getLocation();
-            location.setPitch(playerSender.getLocation().getPitch());
-            location.setYaw(playerSender.getLocation().getYaw());
-            playerSender.teleport(location);
+            this.handler.delete(pos);
 
-            this.plugin.getChatUtils().sendPlayerInfo(playerSender, String.format("Teleporting you to %s%s%s...", Constants.SPECIAL_COLOR, name, Constants.SUCCESS_COLOR));
+            this.plugin.getChatUtils().sendPlayerInfo(playerSender, String.format(
+                    "The %s %s%s%s has been removed !",
+                    this.type.getDisplayName(),
+                    Constants.SPECIAL_COLOR,
+                    name,
+                    Constants.SUCCESS_COLOR
+            ));
 
             return true;
         }
@@ -73,14 +75,16 @@ public abstract class TeleportCommand implements CommandExecutor, TabCompleter {
         if (commandSender instanceof Player) {
             Player playerSender = (Player) commandSender;
 
-            List<String> homeNameList = new ArrayList<>();
+            List<String> nameList = new ArrayList<>();
             List<SavedPosition> list = this.handler.getAllByPlayer(playerSender);
 
             for (SavedPosition pos : list) {
-                homeNameList.add(pos.getName());
+                nameList.add(pos.getName());
             }
-            return homeNameList;
+
+            return nameList;
         }
+
         return Collections.emptyList();
     }
 }
